Day-2
Given an array arr[ ] of n integers and a target value, check if there exists a pair whose sum equals the target. Sorting and Two-Pointer Technique - O(n × log(n)) time and O(1) space. The idea is to use the two-pointer technique but for using the two-pointer technique, the array must be sorted. Once the array is sorted then we can use this approach by keeping one pointer at the beginning (left) and another at the end (right) of the array. 

Check the sum of the elements at these two pointers:

If the sum equals the target, we’ve found the pair.
If the sum is less than the target, move the left pointer to the right to increase the sum.
If the sum is greater than the target, move the right pointer to the left to decrease the sum.